Standardize mkdir and chmod functionality
* Move mkdir and chmod functionality from ffmpeg_utils to path_utils
* Refactor mkdir_recursive and add ensure_path and ensure_dir
* Refactor chmod to be a bit safer and re-use more code
* Add chmod_parent
