+++ include/video/ffmpeg_utils.h
@@ -157,31 +157,4 @@ void jpeg_encoder_cleanup_all(void);
 int ffmpeg_concat_ts_to_mp4(const char **segment_paths, int segment_count,
                             const char *output_path);
 
-/**
- * Create a directory recursively (like mkdir -p)
- * This replaces system("mkdir -p ...") calls
- *
- * @param path Directory path to create
- * @return 0 on success, -1 on error
- */
-int mkdir_recursive(const char *path);
-
-/**
- * Set permissions on a file or directory (like chmod)
- *
- * @param path Path to set permissions on
- * @param mode Permission mode (e.g., 0777)
- * @return 0 on success, -1 on error
- */
-int chmod_path(const char *path, mode_t mode);
-
-/**
- * Recursively set permissions on a directory and its contents (like chmod -R)
- *
- * @param path Directory path to chmod recursively
- * @param mode Permission mode (e.g., 0777)
- * @return 0 on success, -1 on error
- */
-int chmod_recursive(const char *path, mode_t mode);
-
 #endif /* FFMPEG_UTILS_H */
