+++ src/video/ffmpeg_utils.c
@@ -4,8 +4,6 @@
 #include "video/ffmpeg_leak_detector.h"
 #include "video/stream_protocol.h"
 
-#include <sys/stat.h>
-#include <sys/types.h>
 #include <fcntl.h>
 #include <unistd.h>
 #include <errno.h>
@@ -334,145 +332,6 @@ int periodic_ffmpeg_reset(AVFormatContext **input_ctx_ptr, const char *url, int 
 
     log_info("Successfully reset FFmpeg resources");
     return 0;
-}
-
-/**
- * Create a directory recursively (like mkdir -p)
- */
-int mkdir_recursive(const char *path) {
-    if (!path || !*path) {
-        return -1;
-    }
-
-    // Make a mutable copy of the path
-    char *path_copy = strdup(path);
-    if (!path_copy) {
-        log_error("Failed to allocate memory for path copy");
-        return -1;
-    }
-
-    // Iterate through path components and create each directory
-    char *p = path_copy;
-
-    // Skip leading slash for absolute paths
-    if (*p == '/') {
-        p++;
-    }
-
-    while (*p) {
-        // Find next slash
-        while (*p && *p != '/') {
-            p++;
-        }
-
-        char saved = *p;
-        *p = '\0';
-
-        // Create this directory level
-        if (mkdir(path_copy, 0755) != 0 && errno != EEXIST) {
-            log_error("Failed to create directory %s: %s", path_copy, strerror(errno));
-            free(path_copy);
-            return -1;
-        }
-
-        *p = saved;
-        if (*p) {
-            p++;
-        }
-    }
-
-    free(path_copy);
-    return 0;
-}
-
-/**
- * Set permissions on a file or directory (like chmod)
- */
-int chmod_path(const char *path, mode_t mode) {
-    if (!path || !*path) {
-        return -1;
-    }
-
-    if (chmod(path, mode) != 0) {
-        log_warn("Failed to chmod %s: %s", path, strerror(errno));
-        return -1;
-    }
-
-    return 0;
-}
-
-/**
- * Recursively set permissions on a directory and its contents (like chmod -R)
- * Note: This is a simplified version that only sets permissions on the directory itself
- * For full recursive chmod, we would need to walk the directory tree
- */
-int chmod_recursive(const char *path, mode_t mode) {
-    if (!path || !*path) {
-        return -1;
-    }
-
-    // Open as a directory first (O_NOFOLLOW prevents following symlinks, eliminating
-    // the TOCTOU race between the old lstat() check and chmod() on the same path).
-    int fd = open(path, O_RDONLY | O_DIRECTORY | O_NOFOLLOW);
-    bool is_dir = (fd >= 0);
-
-    if (!is_dir) {
-        if (errno == ENOTDIR) {
-            // Not a directory — open as a regular file, still with O_NOFOLLOW
-            fd = open(path, O_RDONLY | O_NOFOLLOW);
-        }
-        if (fd < 0) {
-            if (errno == ELOOP) {
-                // Path is a symlink — skip without error (don't follow)
-                return 0;
-            }
-            log_warn("Failed to open %s: %s", path, strerror(errno));
-            return -1;
-        }
-    }
-
-    // Apply permissions atomically through the open fd (no TOCTOU)
-    if (fchmod(fd, mode) != 0) {
-        log_warn("Failed to fchmod %s: %s", path, strerror(errno));
-        close(fd);
-        return -1;
-    }
-
-    if (!is_dir) {
-        close(fd);
-        return 0;
-    }
-
-    // Directory: use fdopendir so readdir operates on the already-open fd
-    DIR *dir = fdopendir(fd);
-    if (!dir) {
-        log_warn("Failed to fdopendir %s: %s", path, strerror(errno));
-        close(fd);
-        return -1;
-    }
-    // fd is now owned by dir; do not close it separately
-
-    const struct dirent *entry;
-    char full_path[PATH_MAX];
-    int result = 0;
-
-    while ((entry = readdir(dir)) != NULL) {
-        // Skip . and ..
-        if (strcmp(entry->d_name, ".") == 0 || strcmp(entry->d_name, "..") == 0) {
-            continue;
-        }
-
-        snprintf(full_path, sizeof(full_path), "%s/%s", path, entry->d_name);
-
-        // Recursively chmod (opens each entry with O_NOFOLLOW)
-        if (chmod_recursive(full_path, mode) != 0) {
-            result = -1;
-            // Continue processing other entries
-        }
-    }
-
-    closedir(dir); // also closes fd
-    return result;
 }
 
 /**
